# ARROW-7136: [Rust] Added caching to the docker image