--- a/.dockerignore
+++ b/.dockerignore
@@ -48,6 +48,8 @@
 !ruby/red-plasma/Gemfile
 !ruby/red-plasma/lib/plasma/version.rb
 !ruby/red-plasma/red-plasma.gemspec
+!rust/Cargo.toml
+!rust/benchmarks/Cargo.toml
 !rust/arrow/Cargo.toml
 !rust/arrow/benches
 !rust/arrow-flight/Cargo.toml
@@ -55,3 +57,4 @@
 !rust/parquet/build.rs
 !rust/datafusion/Cargo.toml
 !rust/datafusion/benches
+!rust/integration-testing/Cargo.toml