--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -64,6 +64,12 @@ jobs:
       - name: Free Up Disk Space
         shell: bash
         run: ci/scripts/util_cleanup.sh
+      - name: Cache Docker Volumes
+        uses: actions/cache@v1
+        with:
+          path: .docker
+          key: debian-rust-${{ hashFiles('rust/**') }}
+          restore-keys: debian-rust-
       - name: Setup Python
         uses: actions/setup-python@v1
         with: