--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -112,6 +112,12 @@ jobs:
       - name: Fetch Submodules and Tags
         shell: bash
         run: ci/scripts/util_checkout.sh
+      - name: Cache Build Artifacts
+        uses: actions/cache@v1
+        with:
+          path: rust/target
+          key: windows-rust-${{ hashFiles('rust/**') }}
+          restore-keys: windows-rust-
       - name: Build
         shell: bash
         run: ci/scripts/rust_build.sh $(pwd) $(pwd)/build
@@ -144,6 +150,12 @@ jobs:
       - name: Fetch Submodules and Tags
         shell: bash
         run: ci/scripts/util_checkout.sh
+      - name: Cache Build Artifacts
+        uses: actions/cache@v1
+        with:
+          path: rust/target
+          key: macos-rust-${{ hashFiles('rust/**') }}
+          restore-keys: macos-rust-
       - name: Run Clippy
         shell: bash
         run: ci/scripts/rust_lint.sh $(pwd)