--- a/ci/scripts/rust_build.sh
+++ b/ci/scripts/rust_build.sh
@@ -23,7 +23,6 @@ source_dir=${1}/rust
 
 export ARROW_TEST_DATA=${arrow_dir}/testing/data
 export PARQUET_TEST_DATA=${arrow_dir}/cpp/submodules/parquet-testing/data
-export RUSTFLAGS="-D warnings"
 
 # show activated toolchain
 rustup show