--- a/ci/scripts/rust_test.sh
+++ b/ci/scripts/rust_test.sh
@@ -21,7 +21,6 @@ set -ex
 
 arrow_dir=${1}
 source_dir=${1}/rust
-build_dir=${2}/rust
 
 export ARROW_TEST_DATA=${arrow_dir}/testing/data
 export PARQUET_TEST_DATA=${arrow_dir}/cpp/submodules/parquet-testing/data