--- a/ci/scripts/rust_test.sh
+++ b/ci/scripts/rust_test.sh
@@ -25,7 +25,6 @@ build_dir=${2}/rust
 
 export ARROW_TEST_DATA=${arrow_dir}/testing/data
 export PARQUET_TEST_DATA=${arrow_dir}/cpp/submodules/parquet-testing/data
-export CARGO_TARGET_DIR=${build_dir}
 
 pushd ${source_dir}