--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -939,11 +939,10 @@ services:
         arch: ${ARCH}
         rust: ${RUST}
     shm_size: *shm-size
-    environment:
-      CARGO_HOME: /build/cargo
     volumes: *debian-volumes
     command: &rust-command >
       /bin/bash -c "
+        mv /target/ /arrow/rust/ &&
         echo ${RUST} > /arrow/rust/rust-toolchain &&
         /arrow/ci/scripts/rust_build.sh /arrow /build &&
         /arrow/ci/scripts/rust_test.sh /arrow /build"