--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -939,11 +939,13 @@ services:
         arch: ${ARCH}
         rust: ${RUST}
     shm_size: *shm-size
-    volumes: *debian-volumes
+    volumes: &rust-volumes
+      - .:/arrow:delegated
+      - ${DOCKER_VOLUME_DIRECTORY:-.docker}/${ARCH}-debian-${DEBIAN}-rust:/build:delegated
     command: &rust-command >
       /bin/bash -c "
-        mv /target/ /arrow/rust/ &&
         echo ${RUST} > /arrow/rust/rust-toolchain &&
+        if [ ! -d "/build/rust" ]; then mv /rust/dummy-build /build/rust; fi &&
         /arrow/ci/scripts/rust_build.sh /arrow /build &&
         /arrow/ci/scripts/rust_test.sh /arrow /build"