This allow us to build the crate without having to re-build nor download all its dependencies at the expense of a larger (build) image.

Here I am extending the ideas of a non-workspace rust project to a workspace rust project. See [here](https://github.com/rust-lang/cargo/issues/2644) for a _long_ discussion on the issue. Essentially, it is not easy to do this atm with cargo. 

IMO I think that this is worth it, but I need your help to judge it; it incurs some maintenance (duplication of configurations).