@jorgecarleitao 2020-07-20
Some benchmarks:

* Current: 19m with image caching; 21m without image caching
* This PR: ~8m with image caching; 13m without image caching

(~8 assumes 1m download time of the image vs the current 40s, due to bigger image size)

2m:20s are used to build the current layers in both cases

In this PR, 2m40s is used to compile rust dependencies (within docker build). The remaining difference is explained by the second commit on this PR, which was causing a complete cache miss of all built dependencies by one of the executions.

Small drill-down:

* build project (all workspaces + small dependencies): 3m
* build of `cargo check --all-targets --no-default-features`: 40s
* build of `cargo test --features simd`: 1m:30s
* build of examples: ~40s

Overall, I think that there is not so much more to save; it is mostly about building our own code.