@kszucs 2020-07-20
Thanks @jorgecarleitao, it looks like a nice improvement. Currently I'm working on the next release candidate, but I'll take a look at it afterwards.