@kszucs 2020-07-27
Just started to experiment with this change locally. If you don't mind I'm going to push a couple of little changes to better align with the docker-compose setup.