@jorgecarleitao 2020-07-27
LGTM your changes, @kszucs .

> Executed a couple of runs on a beefy machine:
> 
>     * Before my patch: 3m56.235s
> 
>     * Tried without a mounted build volume: 3m42.993s
> 
>     * First run with a mounted build volume : 4m14.134s
> 
>     * First run with a mounted build volume: 1m23.795s

did you run it before this whole PR, just for the sake of sanity checking this whole PR.