@kszucs 2020-07-27
I did, on master the build takes 4m31.312s. The machine has 40 cores and gigabit network, so depending on the dependency-tree's width and the bandwidth the overall improvement might be better on a smaller machine.

It may worth to remove the mounted volume to have better cold build times on the CI. I'm going to add three CI configurations to check the improvements on GHA directly.