@kszucs 2020-07-28
Master - 19m 32s (including docker pull)
No Cache - 13m 28s
Download Cache - 13m 43s
Download and Build Cache - 12m 20s
Download and Build Cache on a Volume - 9m 8s

See build logs [here](https://github.com/apache/arrow/runs/918765051).

The environment variables may not work as expected, but still an improvement over the current master (even if we add the 5 minute build time which we won't need to pay after the docker image gets updated).

I'm going to finalize the PR.